feat: enable lambda deployment with serverless
