+++ index.js
@@ -1,0 +1,5 @@
+const smartApp = require('./smartapp');
+
+module.exports.handle = (event, context, callback) => {
+    smartApp.handleLambdaCallback(event, context, callback);
+};
