+++ test/weather.spec.js
@@ -1,6 +1,6 @@
 const test = require('ava');
 const colorType = require('../constants/colors');
-const weather = require('./weather');
+const weather = require('../lib/weather');
 const forecast = require('../config/forecast-test.json');
 
 test('cloudy first chunk snow second chunk should be purple', t => {
